pages(downloads): refactor into components
